refactor(sync): split engine.go below 1000 lines
Move push, pull, apply, and persist helpers into sibling files
in package sync. No behavior change.

Related: #652
